test(harness): move live Anthropic chat tests to chat_live_anthropic suite
From test_anthropic_completion.py per audit 5/8c: TestAnthropicCompletion
(BaseLLMChatTest + BaseAnthropicChatTest), the live tool_use /
structured_output / citations / thinking / websearch / text_editor /
mcp_server / prefix_prompt / streaming / responses-bridge / replay tests,
test_just_system_message, test_claude_tool_use_with_anthropic_acreate.
These are provider truth re-verified every 4h; largely overlapping the
compat matrix rows, so dedupe against tests/claude_code/manifest.yaml when
wiring cells. The transform/parse goldens (headers, tool helpers, json-mode,
metadata filters, stop sequences, raw-request body captures) stay behind as
unit tests beside v2.
